let's start another attempt with restructing the code
